+++ tests/test_log_progress.py
@@ -13,10 +13,11 @@ def test_extra(extra, caplog, monkeypatch):
     )
 
     user_extra = {"extra0": 0, "extra1": 1, "mutable": "a"}
-    for i, _ in enumerate(
-        log_split_progress(split(["2022-1", "2022-2", "2022-3"], after="1d"), extra=user_extra),
-        start=1,
-    ):
+
+    progress = log_split_progress(split(["2022-1", "2022-2", "2022-3"], after="1d"), extra=user_extra)
+    assert len(progress) == len(progress.splits)
+
+    for i, _ in enumerate(progress, start=1):
         record = caplog.records[-1]
         must_contain = extra.format(**user_extra)
         assert record.message.startswith("Begin fold")